--- a/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/FlattenRunner.java
+++ b/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/FlattenRunner.java
@@ -22,13 +22,9 @@ import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.I
 import com.google.auto.service.AutoService;
 import java.io.IOException;
 import java.util.Map;
-import org.apache.beam.model.pipeline.v1.RunnerApi.Components;
-import org.apache.beam.model.pipeline.v1.RunnerApi.PCollection;
 import org.apache.beam.runners.core.construction.PTransformTranslation;
-import org.apache.beam.runners.core.construction.RehydratedComponents;
 import org.apache.beam.sdk.fn.data.FnDataReceiver;
 import org.apache.beam.sdk.util.WindowedValue;
-import org.apache.beam.sdk.util.WindowedValue.WindowedValueCoder;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap;
 
 /** Executes flatten PTransforms. */
@@ -50,41 +46,16 @@ public class FlattenRunner<InputT> {
   static class Factory<InputT> implements PTransformRunnerFactory<FlattenRunner<InputT>> {
     @Override
     public FlattenRunner<InputT> createRunnerForPTransform(Context context) throws IOException {
-
       // Give each input a MultiplexingFnDataReceiver to all outputs of the flatten.
       String output = getOnlyElement(context.getPTransform().getOutputsMap().values());
       FnDataReceiver<WindowedValue<?>> receiver = context.getPCollectionConsumer(output);
 
-      RehydratedComponents components =
-          RehydratedComponents.forComponents(
-              Components.newBuilder().putAllCoders(context.getCoders()).build());
-
       FlattenRunner<InputT> runner = new FlattenRunner<>();
       for (String pCollectionId : context.getPTransform().getInputsMap().values()) {
-        context.addPCollectionConsumer(
-            pCollectionId,
-            (FnDataReceiver) receiver,
-            getValueCoder(components, context.getPCollections(), pCollectionId));
+        context.addPCollectionConsumer(pCollectionId, (FnDataReceiver) receiver);
       }
 
       return runner;
-    }
-
-    private org.apache.beam.sdk.coders.Coder<?> getValueCoder(
-        RehydratedComponents components,
-        Map<String, PCollection> pCollections,
-        String pCollectionId)
-        throws IOException {
-      if (!pCollections.containsKey(pCollectionId)) {
-        throw new IllegalArgumentException(
-            String.format("Missing PCollection for id: %s", pCollectionId));
-      }
-      org.apache.beam.sdk.coders.Coder<?> coder =
-          components.getCoder(pCollections.get(pCollectionId).getCoderId());
-      if (coder instanceof WindowedValueCoder) {
-        coder = ((WindowedValueCoder<InputT>) coder).getValueCoder();
-      }
-      return coder;
     }
   }
 }