@lukecwik 2022-06-22
![tiny bundle](https://user-images.githubusercontent.com/10078956/175115408-882e810f-07ca-4d94-b231-deb2a177ad66.png)
![large bundle](https://user-images.githubusercontent.com/10078956/175115423-76f92110-1cdb-40ca-9e0c-cd69e992d968.png)