@lukecwik 2022-06-22
I think that reviewing the interaction with the ProcessBundleHandler and the new metrics class E2E will provide the best context and review quality vs splitting this up. I was planning on throwing in more changes but thought that this was a reasonably sized change. See the follow-up in https://github.com/apache/beam/pull/22005